+++ spec/lib/sequel_rails/railties/database_rake_spec.rb
@@ -5,20 +5,22 @@ describe "Database rake tasks" do
   let(:app) { Combustion::Application }
   let(:app_root) { app.root }
 
+  around do |example|
+    begin
+      FileUtils.rm schema if File.exists? schema
+      example.run
+    ensure
+      FileUtils.rm schema if File.exists? schema
+    end
+  end
+
   describe "db:schema:dump" do
     let(:schema) { "#{app_root}/db/schema.rb" }
-    around do |example|
-      FileUtils.cp schema, "#{schema}.orig"
-      begin
-        example.run
-      ensure
-        FileUtils.mv "#{schema}.orig", schema
-      end
-    end
 
     it "dumps the schema in 'db/schema.rb'" do
       Dir.chdir app_root do
-        expect{`rake db:schema:dump`}.to change{File.mtime schema}
+        `rake db:schema:dump`
+        File.exists?(schema).should be_true
       end
     end
 
@@ -26,10 +28,12 @@ describe "Database rake tasks" do
       Dir.chdir app_root do
         `rake db:migrate`
         `rake db:schema:dump`
+        sql = Sequel::Model.db.from(:schema_migrations).
+          insert_sql(:filename => "1273253849_add_twitter_handle_to_users.rb")
         File.read(schema).should include <<-EOS
 Sequel.migration do
   change do
-    self << "INSERT INTO \\\"schema_migrations\\\" (\\\"filename\\\") VALUES ('1273253849_add_twitter_handle_to_users.rb')"
+    self << #{sql.inspect}
   end
 end
 EOS
@@ -39,17 +43,9 @@ EOS
 
   describe "db:structure:dump" do
     let(:schema) { "#{app_root}/db/structure.sql" }
-    around do |example|
-      begin
-        example.run
-      ensure
-        FileUtils.rm schema
-      end
-    end
 
     it "dumps the schema in 'db/structure.sql'" do
       Dir.chdir app_root do
-        File.exists?(schema).should_not be_true
         `rake db:structure:dump`
         File.exists?(schema).should be_true
       end
@@ -60,7 +56,8 @@ EOS
         `rake db:migrate`
         `rake db:structure:dump`
 
-        sql = "INSERT INTO \\\"schema_migrations\\\" (\\\"filename\\\") VALUES ('1273253849_add_twitter_handle_to_users.rb')"
+        sql = Sequel::Model.db.from(:schema_migrations).
+          insert_sql(:filename => "1273253849_add_twitter_handle_to_users.rb")
         File.read(schema).should include sql
       end
     end