+++ History.md
@@ -1,6 +1,8 @@
 0.5.1.dev
 =========
 
+* Allow setting the search path in app config (Rafał Rzepecki) [#36]
+
 0.5.0 (2013-07-08)
 ==================